+++ app/src/main/kotlin/com/pitchedapps/frost/db/CookiesDb.kt
@@ -84,7 +84,8 @@ object CookiesDb {
 @Parcelize
 @Table(database = CookiesDb::class, allFields = true, primaryKeyConflict = ConflictAction.REPLACE)
 data class CookieModel(
-    @PrimaryKey var id: Long = -1L, var name: String? = null,
+    @PrimaryKey var id: Long = -1L,
+    var name: String? = null,
     var cookie: String? = null
 ) :
     BaseModel(), Parcelable {