+++ app/src/main/kotlin/com/pitchedapps/frost/FrostApp.kt
@@ -35,7 +35,6 @@ import com.pitchedapps.frost.db.FbTabsDb
 import com.pitchedapps.frost.db.FrostDatabase
 import com.pitchedapps.frost.db.NotificationDb
 import com.pitchedapps.frost.glide.GlideApp
-import com.pitchedapps.frost.services.scheduleNotifications
 import com.pitchedapps.frost.services.scheduleNotificationsFromPrefs
 import com.pitchedapps.frost.services.setupNotificationChannels
 import com.pitchedapps.frost.utils.BuildUtils