+++ app/src/main/kotlin/com/pitchedapps/frost/utils/Prefs.kt
@@ -191,7 +191,9 @@ object Prefs : KPref() {
 
     var showCreateFab: Boolean by kpref("show_create_fab", true)
 
-    var webOnly: Boolean by kpref("web_only", false)
+    var authRequests: Boolean by kpref("web_requests", false)
+
+    var nativeUi: Boolean by kpref("native_ui", true)
 
     inline val mainActivityLayout: MainActivityLayout
         get() = MainActivityLayout(mainActivityLayoutType)