+++ docs/Changelog.md
@@ -6,6 +6,8 @@
 * Disable bugsnag completely when opting out of analytics
 * Filter urls before sending to other apps
 * Allow hiding main fab (see settings > newsfeed)
+* Add some experimental options to debug login problems (settings > experimental)
+* Enforce desktop user agent for now
 
 ## v2.3.1
 * Hide all story panels if enabled